Use deafness global setting

+++ server/src/global_settings.rs
@@ -29,7 +29,7 @@ impl GlobalSettings {
 pub struct BattleSettings {
     pub fatigue_start: u32,
     pub deafness_start: u32,
-    pub deafness_per_turn: f32,
+    pub deafness_per_turn: f64,
 }
 
 #[derive(SpacetimeType)]
